+++ src/com/atteo/jello/space/AppendOnlyCache.java
@@ -2,7 +2,7 @@ package com.atteo.jello.space;
 
 public interface AppendOnlyCache {
 	public boolean isEmpty();
-	public long getBestId(int freeSpace);
-	public int getFreeSpace(long id);
-	public void update(long id, int freeSpace);
+	public int getBestId(short freeSpace);
+	public short getFreeSpace(int id);
+	public void update(int id, short freeSpace);
 }
